+++ docs/changelog.md
@@ -10,7 +10,9 @@ pageflow_next_text: Contributing
 
 ## develop branch
 
-Nothing yet.
+### New
+
+* Added support for a default call stack filter
 
 ## v1.2016061901